requirements.txt
wandb==0.18.1
fastapi==0.115.0
torch==2.4.1
python-dotenv==1.0.1
opencv-python>=4.5.5.64
mediapipe>=0.8.9.1
tqdm>=4.64.0
librosa>=0.10.2.post1
moviepy>=1.0.3
python-multipart>=0.0.5